Split gmp utils into several js modules
