+++ ng/src/gmp/utils/__tests__/calc.js
@@ -1,0 +1,95 @@
+/* Greenbone Security Assistant
+ *
+ * Authors:
+ * Björn Ricks <bjoern.ricks@greenbone.net>
+ *
+ * Copyright:
+ * Copyright (C) 2018 Greenbone Networks GmbH
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+import {
+  avg,
+  sum,
+} from '../calc';
+
+describe('sum function tests', () => {
+
+  test('should calculate the array sum', () => {
+    expect(sum([1, 2, 3])).toBe(6);
+  });
+
+  test('should calculate sum from object props', () => {
+    const array = [
+      {value: 1},
+      {value: 2},
+      {value: 3},
+    ];
+    expect(sum(array, entry => entry.value)).toBe(6);
+  });
+
+  test('should expect zero for undefined values', () => {
+    expect(sum([1, 2, undefined, 3])).toBe(6);
+
+    const array = [
+      {value: 1},
+      {},
+      {value: 2},
+      {value: 3},
+    ];
+    expect(sum(array, entry => entry.value)).toBe(6);
+  });
+
+  test('should sum empty and undefined to zero', () => {
+    expect(sum()).toBe(0);
+    expect(sum([])).toBe(0);
+  });
+});
+
+describe('avg function tests', () => {
+
+  test('should calculate the array average', () => {
+    expect(avg([3, 3, 3])).toBe(3);
+  });
+
+  test('should calculate average from object props', () => {
+    const array = [
+      {value: 3},
+      {value: 3},
+      {value: 3},
+    ];
+    expect(avg(array, entry => entry.value)).toBe(3);
+  });
+
+  test('should expect zero for undefined values', () => {
+    expect(avg([5, undefined, 4])).toBe(3);
+
+    const array = [
+      {value: 5},
+      {},
+      {value: 4},
+    ];
+    expect(avg(array, entry => entry.value)).toBe(3);
+  });
+
+  test('should return 0 for empty array', () => {
+    expect(avg([])).toBe(0);
+    expect(avg([], item => item.value)).toBe(0);
+    expect(avg()).toBe(0);
+    expect(avg(undefined, item => item.value)).toBe(0);
+  });
+});
+
+// vim: set ts=2 sw=2 tw=80:
