+++ ng/src/gmp/utils/__tests__/object.js
@@ -1,0 +1,56 @@
+/* Greenbone Security Assistant
+ *
+ * Authors:
+ * Björn Ricks <bjoern.ricks@greenbone.net>
+ *
+ * Copyright:
+ * Copyright (C) 2018 Greenbone Networks GmbH
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+import {
+  exclude,
+  exclude_object_props,
+} from '../object';
+
+describe('exclude function test', () => {
+  test('exclude object property', () => {
+    const obj = {
+      foo: 1,
+      bar: 2,
+    };
+    const result = exclude(obj, prop => prop === 'foo');
+
+    expect(result.foo).toBeUndefined();
+    expect(result.bar).toBe(2);
+  });
+});
+
+describe('exclude_object_props function test', () => {
+  test('exclude object properties', () => {
+    const obj = {
+      foo: 1,
+      bar: 2,
+      abc: 3,
+    };
+    const result = exclude_object_props(obj, ['foo', 'bar']);
+
+    expect(result.foo).toBeUndefined();
+    expect(result.bar).toBeUndefined();
+    expect(result.abc).toBe(3);
+  });
+});
+
+// vim: set ts=2 sw=2 tw=80:
